introduce TaskStatusService abstraction
